Add Option to Control Display of "out:" Box in LineZoneAnnotator
Summary:
Introduced a `display_out` parameter in the `LineZoneAnnotator` class to provide users the flexibility to decide whether to render the "out:" box or not.

Changes:
- Added `display_out` attribute with a default value of `True` in the initializer.
- Conditionally wrapped the rendering logic for the "out:" box based on the value of `display_out`.

Impact:
This change is backward compatible, as the default behavior remains unchanged. However, users now have the option to hide the "out:" box if desired.