Improved the mask_to_polygons function to handle masks that have exclusions and Masks that do not align
Example:
Detections of Doughnut: Even if the Model correctly detects the Doughnut, the Polygon always was only the outer shape. With this change the hole in the middle is excluded from the Polygon
Person behind a table: When the Model identifies the Person correctly, the Part above and below the Table would have been two seaperate Polygons. With this change, it would be one Polygon connected.